fix(nodes): resolve node ids to content items
